requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
# --- PyTorch Installation for Mac M1/CPU ---
torch
torchvision
torchaudio

# --- Hugging Face Transformers and dependencies ---
tritonclient[http]
evaluate
scikit-learn
transformers
accelerate
sentencepiece

# --- Other ---
requests
pytest
pytest-cov
requests-mock
httpx
pytest
pytest-cov
requests-mock
httpx
